# WIP: Optimization of einsum in shc_B_H