I optimized `shc_B_H` by replacing `einsum` with matrix multiplication, transposing and reshaping. This improves the performance of opt_SHCqiao calculation.

Optimization of `einsum` reduces the time taken in `shc_B_H` is from 243s to 82s. After the optimization, the bottleneck is `_R_to_k_H` inside `shc_B_H`  (~50% of total time). See details below.

I checked with bulk Pt that the calculation result does not change after optimization.

This optimization was discussed in https://github.com/stepan-tsirkin/wannier-berri/pull/22#discussion_r532286534
I tried 1) using `einsum_path` and 2) factoring out the 'k' index from the einsum call (but still using einsum inside the loop). But, both did not give any speedup.

Dear @migueldiascosta, could you (or the user, Truman Ng Yu - I could not find the GitHub id) try this MR and double check its performance and correctness?

### Summary of the performance test
I used `num_wann=88`, `number of R points = 1163`, `SHCqiao=True`, `len(Efermi)=1`, `len(omega)=1`, `NKdiv=[1 1 1]`, and `NKFFT=[10 10 10]`. I used `kernprof` to profile the code.

If there are ~100 Efermi or omega, `kubo_sum_elements` becomes the bottleneck.
Note that `kubo_sum_elements` scales as O(N_W^2) while `_R_to_k_H` and `_shc_B_H_einsum_opt` scales as O(N_W^3). So, `shc_B_H` becomes more important when `N_W` is larger.

* Before/after optimization, `shc_B_H` takes 90%/76% of time.
* Optimization of `einsum` reduces the time taken in `shc_B_H` is from 243s to 82s.
* For the `einsum` lines, the speedup is from 180s to 20s.
* After optimization, `_R_to_k_H` of `SH_R`, `SR_R` and `SHR_R` takes ~57% of the total time. Optimized einsum takes ~17%. Calculation of `A_H` also takes ~17%.
* Inside `_shc_B_H_einsum_opt`, most (~90%) of the time is spent in the actual matrix multiplication.