@stepan-tsirkin 2021-02-08
Thank you Jaemo for amother optimization.  It would be good if someone tests it on more examples before merge. 
We really need to implement automated testing.. 

So, you say that `_R_to_k_H` in `SH_R` , `SR_R` and `SHR_R` takes more than half of the time. But it is also used to get Hamiltonian, get `S_H` and `A_H` (which consists of `A_Hbar` and `D_H`, the latter depends on `V_H`)  Does it mean that all together this routinne takes a vast majority of the time? If so -  that is the new bottleneck. It consists of Foutier transform and matrix rotations, but I guess for large number of Wannier functions the rotations become dominant, because of the numwann^3 scaling that you mentioned.

It would be nice to find a way to optimize this matrix rotation. As an option we may re-write that piece of code in C or Fortran, with calls to Blas. 

Another thing: I dod some attempts to make parallelization inside Data_K class, for case when parallelisation over K-point6s is impossible. Probably  that implementation is slowing down. Maybe cleaning the code from those things cvan help. Previously there was a restriction on the size of minimal FFT-grid and also the memory usage for some quantities was huge. Now we can use arbitrary small FFT grids, and also memory usage was optimized. Therefor probably doing _rotations in parallel is not needed anymore.