@stepan-tsirkin 2021-02-08
> I cannot agree more.. Do you have some plans for this?
Not really, I have never done this myself.  do you have any experience?  it is mainly a question  how to choose a proper way of doing it..
> Doesn't np.dot() (and @) call BLAS functions?

Yes, they do, I think. Not sure if we can get much  from using C.fortran. Basically, the overhead of numpy is in crerating and destroying arrays, as well as transpose, but for big matrices this overhead is actually negligible.  Could be that we reached the limit here, but not sure.

>I am curious how this is done.
Also, regarding performance, for a fixed N_k(=N_FFT * N_Kdiv), if one uses a smaller N_FFT, the total time spent in FFT is longer (because FFT scales as O(N_R log N_R)), but with the benefit of having larger NKdiv which is good for parallelization. Is this correct?

It is done easily, as I realized there is no problem. If two R-vectors fall into same place in FFT box - they just add up contributions, and the final result stays the same. Yes, for very small FFT grids, the preformance will go down, so it is a compromise between efficient FFT and efficient parallelisation, and also demand of memory. Usually optimal FFT is same as abinitio grid, but one can go smaller as a compromise, as long as FFT does not become the heaviest part again (like in postw90).

Actually parallelization is another problem. We are still bound to a single node.  We need a better solution than just `multiprocessing` module. THer are some other solutions to study:
https://wiki.python.org/moin/ParallelProcessing