@migueldiascosta 2021-02-08
> This optimization was discussed in [#22 (comment)](https://github.com/stepan-tsirkin/wannier-berri/pull/22#discussion_r532286534)
> I tried 1) using `einsum_path` and 2) factoring out the 'k' index from the einsum call (but still using einsum inside the loop). But, both did not give any speedup.

@jaemolihm Once you factor out the 'k' index, the `einsum` calls inside the loop will also use BLAS functions, so that would already be much better than before and similar to what you are doing now (still, you're right, the way you do it now is even faster, by ~10% in the example I'm using)