@stepan-tsirkin 2021-02-08
@migueldiascosta ,  so you checked and the results did not change after optimization, right? 
When you confirm that, I will merge