@jaemolihm 2021-02-08
> > This optimization was discussed in [#22 (comment)](https://github.com/stepan-tsirkin/wannier-berri/pull/22#discussion_r532286534)
> > I tried 1) using `einsum_path` and 2) factoring out the 'k' index from the einsum call (but still using einsum inside the loop). But, both did not give any speedup.
> 
> @jaemolihm Once you factor out the 'k' index, the `einsum` calls inside the loop will also use BLAS functions, so that would already be much better than before (~10x, in the example I'm using) and similar to what you are doing now (still, you're right, the way you do it now is even faster than that, an additional ~1.1x in the example I'm using)

Hi @migueldiascosta, thank you for the comments.

I am a bit confused. I ran the following script that implements what you suggested (or my understanding of it): https://gist.github.com/jaemolihm/9e0b8ae2852e4b397489586fc068ff99

The function is
```python3
def loop_plus_einsum(C, A, B):
    # Optimized version of C += np.einsum('knlc,klma->knmac', A, B). Used in shc_B_H.
    for ik in range(nk):
        C[ik] = np.einsum('nlc,lma->nmac', A[ik], B[ik])
```

The output is as follows.
```
nk = 100, nw = 50, repetition = 5
einsum        :  6.747 sec
loop + einsum :  6.774 sec
matmul        :  0.240 sec
```

So, the loop + einsum version did not give any speedup. Is the function above what you suggested, or have I misunderstood you?