@migueldiascosta 2021-02-08
@jaemolihm try passing `optimize=True` to all the `einsum` calls in that script - it should have no effect on "naive einsum", but it should make "loop + einsum" almost as fast as "matmul"