@stepan-tsirkin 2021-02-08
Thank you @migueldiascosta , I did not know about the `optimize` option. It is useful, although not as good as expanding sums explicitly:

```
In [30]: c.shape                                                                                 
Out[30]: (10000, 20, 20)

In [31]: %timeit np.einsum('klm,kmn,knp->klp',a,b,c)                                             
2.36 s ± 5.15 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [32]: %timeit np.einsum('klm,kmn,knp->klp',a,b,c,optimize=True)                               
210 ms ± 881 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [33]: %timeit np.array([x @ y @ z for x,y,z in zip(a,b,c)])                                   
64 ms ± 182 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

```