@stepan-tsirkin 2021-02-08
@migueldiascosta  , do you think it is useful and harmless to put optimize=True to all calls of einsum in the code?
Or is there areason why this option is not set to default in numpy yet?