@migueldiascosta 2021-02-08
@stepan-tsirkin for very small matrices the overhead can be high, e.g. see https://github.com/numpy/numpy/issues/11714 and comments