@migueldiascosta 2021-02-09
> @migueldiascosta , so you checked and the results did not change after optimization, right?
> When you confirm that, I will merge

@stepan-tsirkin apart from the small numerical differences (~10^-10) that come from using an optimized BLAS library, I get the same results